factorise reducer
