+++ src/reducer/ProductReducer.jsx
@@ -33,5 +33,5 @@ const reducer = (state, action) => {
       return state;
   }
 };
- 
-export {reducer};
+
+export { reducer };
